business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SetCTCSSTone.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SetCTCSSTone implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+C ");
        messageString.append(command.getParameter("CTCSS Tone"));
        messageString.append("\n");
        return messageString;
    }
}
